Recoded the calculation for proportional #experimental
Very experimental code for calculating the proportional. If it works, i believe it will be far more accurate and reliable than previously, and some others teams.
Made some changes to the timestep, kp and kp values. 
New method for calculating the error value, is experimental but should be implemented into the currently tested code whether my new code works or not.

//Reason for the experimental\\
After all the stuff we have done in lectures and labs, i decided that i would try write some code that better reflected trying to ignore noise and gain more accurate and reliable readings.
Starts off with a for loop that runs 10 times, each time it takes a picture and then creates a total value for each pixel for each photo.
Takes the average of each pixel, this will dramatically help with reducing noise and makes the pixel information more reliable.
it then checks if each pixel average is above 127, this is half brightness and im assuming that black and white will be significantly lower or higher than this value respectively.
Other than that the code is unchanged.