new line finding method
robot now stops when it loses the line, and rotates until the line is found again. When it detects the line, we rely on our lineFollower to line the robot up again. Could be more elegantly executed, as there is some duplicate code - the relevant part of lineFollow is just copy-pasted into turnAround. 
I think this will work for dead-ends and T junctions. No need to reverse as the camera should be sufficiently removed from the pivot point to catch the line as it turns around.
It won't work very well if it picks up false positives for white pixels, so hopefully this is tuned correctly.  Also our lineFollower may not be reliable enough to correct based on one white pixel. If it isn't, we can add a sleep to keep it rotating for a bit longer so it has more line to follow.